+++ noodles-vcf/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+  * vcf/record/samples: Add series selector (`Samples::select`).
+
 ## 0.53.0 - 2024-04-11
 
 ### Changed